Fix warning

+++ app.py
@@ -168,7 +168,7 @@ def _search_for_results(query: str) -> pd.DataFrame:
     )
 
     df_out["score"] = (
-        df_out["score"].fillna(0)
+        df_out["score"].astype(float).fillna(0)
         + df_out["name"].apply(_f_score_in_name) * 10
         + df_out["organisation"].apply(_f_score_in_name) * 10
     )
